starting

+++ src/main/java/com/codebrew/repository/UserEventRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserEventRepo extends JpaRepository<UserEvent, Long> {
 
-    UserEvent findUserEventById(Long id);
+   UserEvent findUserEventById(Integer id); 
 
 }
